Delete SeleniumTest project Add NUnit project
