+++ SeleniumTest.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30011.22
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SeleniumTest", "SeleniumTest\SeleniumTest.csproj", "{5E87E6D5-941E-44BA-BF41-7D6862ED5DA8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SeleniumTests", "SeleniumTests\SeleniumTests.csproj", "{B1FE2E44-9110-4E75-8030-3FDCB98B150D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +11,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5E87E6D5-941E-44BA-BF41-7D6862ED5DA8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5E87E6D5-941E-44BA-BF41-7D6862ED5DA8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5E87E6D5-941E-44BA-BF41-7D6862ED5DA8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5E87E6D5-941E-44BA-BF41-7D6862ED5DA8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B1FE2E44-9110-4E75-8030-3FDCB98B150D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B1FE2E44-9110-4E75-8030-3FDCB98B150D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B1FE2E44-9110-4E75-8030-3FDCB98B150D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B1FE2E44-9110-4E75-8030-3FDCB98B150D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
