--- SeleniumTest/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="NUnit" version="3.12.0" targetFramework="net472" />
-  <package id="Selenium.Chrome.WebDriver" version="81.0.0" targetFramework="net472" />
-  <package id="Selenium.WebDriver" version="3.141.0" targetFramework="net472" />
-</packages>
